requirements.txt
Pillow>=2.6.1
RPi.GPIO>=0.6.3
spidev>=3.0
